+++ src/components/Home.js
@@ -157,6 +157,7 @@ const UIComponent = ({ type }) => {
           name="Mehul Lakhanpal"
           email={"mehullakhanpal@gmail.com"}
           url="https://lh3.googleusercontent.com/a-/AOh14Gj3jT11sey95Pv4J12bgoIjqzc9IW0mNiZUjh7o0Q=s96-c"
+          onItemClick={console.log}
         />
       );
     case "ICON":