+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix: Changed to 'source-map'
 - Fix: Added size prop for Profile Avatar
 - Fix: Drop position for 'Dropdown'
+- Fix: Add props for 'Dropdown'
 
 ### v3.0.0